backfill-tool/requirements.txt
# AT Protocol Firehose Backfill Tool - Python Dependencies

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AT Protocol / WebSocket
websockets==12.0
cbor2==5.6.2
dag-cbor==0.3.3
multiformats==0.3.1.post4

# Async/Concurrency
aiohttp==3.9.3
asyncio==3.4.3

# CLI & Progress
click==8.1.7
rich==13.7.0
python-dotenv==1.0.1

# Utilities
python-dateutil==2.8.2
pydantic==2.6.1
